requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.8.0
appscript==1.3.0
backoff==2.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
emoji==2.14.1
eval_type_backport==0.2.2
filelock==3.17.0
filetype==1.2.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Limiter==3.10.1
Flask-SQLAlchemy==3.1.1
fsspec==2025.2.0
ftfy==6.3.1
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
langdetect==1.0.9
limits==4.0.1
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
olefile==0.47
openai==1.64.0
opencv-python==4.11.0.86
ordered-set==4.1.0
packaging==24.2
pdf2image==1.17.0
pillow==11.1.0
psutil==7.0.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
Pygments==2.19.1
pypdf==5.3.0
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
PyYAML==6.0.2
RapidFuzz==3.12.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
safetensors==0.5.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
surya-ocr==0.4.2
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.9.0
tokenizers==0.15.2
tqdm==4.67.1
transformers==4.36.2
typing-inspect==0.9.0
typing_extensions==4.12.2
unstructured==0.16.23
unstructured-client==0.30.4
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
